// frontend/src/client/models/ProjectOut.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ProjectOut = {
    name: string;
    description?: (string | null);
    data_dir: string;
    project_id: number;
    task_category_id?: (number | null);
    created: string;
    modified: string;
};